// src/api/boce-wac-project-ambulance-wl/api.ts
export interface AppointmentsList {
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'date': string;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'estimatedStart': string;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'estimatedEnd': string;
    /**
     * 
     * @type {boolean}
     * @memberof AppointmentsList
     */
    'patientAppointed': boolean;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'condition': string;
    /**
     * 
     * @type {string}
     * @memberof AppointmentsList
     */
    'doctorNote': string;
}